SG-13614: Fails properly when Jekyll encounters a problem with the site's contents (#30)
The config change causes Jekyll build to stop what it's doing when it runs into errors and raise a proper exception. However, that does NOT cause the bundle exec jekyll build call to return non-zero. The upside is that it DOES stop Jekyll from writing the index.html for the site it's building, which means we can check to see if that exists to determine whether the build succeeded or not.